refactor(xapiri): extract dashboard_msgs.go — tea.Msg types + waitForCostRowCmd
Part 1 of ADR 0014 dashboard split. Moves all tea.Msg type definitions and
waitForCostRowCmd out of the 4516-line monolith into dashboard_msgs.go.
Zero behaviour change; package-internal symbols only.

Closes #179
